Voter model
Adds the voter model to the suite of synthetic-data-generating models.

Some notes for reviewers:

1. I use the in-degree normalized columns of the adjacency matrix as
the probabilities, following the logic that, for directed networks,
looking at the alter nodes that are incident on ego makes more sense
than looking at alter nodes upon which ego is incident. I'm open to
argument on this.
2. My understanding is that one time step in the voter model is usually
one interaction. Since this would not make for a lot of dynamics, I
instead defined a time step as one interaction for each node. For
consistency with the original definition, the model updates
asynchronously within a time step.
3. Brennan and I double-checked how `transitions` was constructed, but
in general I could always used my intuitions about columns and rows in
Numpy checked. If something's wrong in the implementation, it's probably
in selecting a row when I meant a column or vice versa.
4. This seemed like the most parsimonious way to add noise to prevent
the model from converging rapidly. Open to suggests on other ways to
implement noise.