Address self-loops when thresholding
Inspired by the discussion in #54. When thresholding on a quantile or average
degree, the diagonal of the weights matrix was being used in the calculation of
the quantile or average degree, even when those self-loops were destined to be
removed by the `create_graph` function. This PR adds a `remove_self_loops`
parameter, passed through `kwargs`, to zero out the weights matrix when
appropriate.