Threshold inverse p-values in CCM
Modify the CCM reconstructor to use the `threshold` function. It does this by
creating a matrix of one minus the p-values, to have similar behavior to other
weight matrices (i.e., maintaining that higher values are more import elements).
To maintain consistency with its previous default behavior, the `threshold_type`
is set to `'range'` and a default `cutoffs` is provided, corresponding to the
95% alpha level of significance used before.

Resolves #182.