Generate PyPI description and long_description from README
Resolves #233. The PyPI description was sparse because we were not defining
`long_description`. Instead of writing a new `long_description`, we now use the
first section of the README to generate both `description` (the title of the
section) and `long_description`. As a side effect (we weren't sure if Markdown
was allowed in `description`), we strip the extraneous code backticks from
"netrd" in the README, and in the docs `index.rst` for consistency. We're going
to push this ASAP and bump to 1.0.1 to make sure the change takes on PyPI.