make NBD deterministic and include in test coverage
Initializes the call to `sp.sparse.linalg.eigs` with a dummy vector of 1/Ns so
that the result is deterministic for a graph. This allows us to include NBD
back in the test suite; however, there is still some minor numerical inaccuracy,
so we change `dist == 0.0` to `np.isclose(dist, 0.0)` to account for this.