Use proper normalization in Hamming (#250)
Harrison pointed out in a comment on our paper that our Hamming implementation
has an implicit $N^2$ instead of $N(N-1)$ normalization, so it's wrong for
graphs without selfloops. This corrects that, similar to #242.

A couple of notes:

1. I think this could be a little cleaner; the fact that `np.triu_indices()` et
al return 2-tuples cramped my style a bit.
2. The fact that this and #242 exist raise concern that this normalization issue
may be present elsewhere. Perhaps we should open a checklist issue, like we have
for #245?
3. I have not applied the same correction to `HammingIpsenMikhailov`, on the
grounds that: (i) it's sufficiently different from regular `Hamming` to consider
separately, and (ii) it probably deserves a more thorough cleanup.