Implement D-Measure
Resolves #253. This is a heavily refactored version of @jkbren's initial
translation of the canonical R code to Python. It produces basically the same
output, including in the intermediate steps; I would attribute the relatively
small differences in final distances to differences in precision, floating-point
error, etc.

One thing that might be problematic: I'm not certain it's going to play nicely
with graphs of difference sizes right now, which is why I'd like to have test
coverage for that. It _should_ be fine, I think. There's probably a little more
refactoring to be done in the `alpha_jsd` function, replacing the fairly manual
padding done there with a function, as in @leotrs's implementation.

Docs don't give the full Eq. 2 from the paper but hopefully give enough of the
gist, and the reference to the paper is in there, of course.