Explicitly test for and handle weighted input  (#276)
* version bump

* Explicitly test for and handle weighted input

Resolves #270. This commit has a large footprint and I'm not sure it's the best
solution for the problem, but it is _a_ solution. The core changes are the
addition of an `ensure_unweighted()` function with behavior analogous to
`ensure_undirected()`, and a new test, `test_weighted_input()`. That test
compares two isomorphic graphs with different weights, and asserts that one of
two things happens:

1. the distance function does not return zero, or
2. the distance function does return zero **and** emits a `RuntimeWarning`
generated by `ensure_unweighted()`.

The downside of this approach is in the fact that all distances that do not
support weighted graphs must add this function, which means a lot of repeated
code and increase possibility of users encountering warnings. To the first, a
decorator would be a natural next step. To the second, a solution would be
subclassing, maybe?

* autoformatter

* streamline ensure_unweighted definition