Bring threshold inside BaseReconstructor class
See #271.

This is presented as a proposal; it's **not ready for merging**. (The tests pass
but a ton of doc changes are missing.)

This consolidates a lot of the thresholding and graph creation logic that exists
in individual reconstructors and in the utilities submodule within the
`BaseReconstructor` class. The advantages of such an approach are as follows:

1. **Minimizes code repetition.** Each of the reconstructors (with one or two
exceptions) currently imports `create_graph()` and `threshold()` and calls those
functions just prior to returning the graph. By moving this functionality within
the base class and using inheritance, there's less duplication and upkeep.
2. **(Mostly) gets rid of kwargs.** The use of `kwargs` within the
reconstruction and thresholding logic has always been extremely messy and
error-prone. Because we no longer need to pass in `threshold_type` and
threshold-specific arguments to `Reconstructor.fit()`, we don't need to capture
`kwargs` in that method call. The general-purpose `Reconstructor.threshold()`
still uses `kwargs` to dispatch threshold method-specific arguments, but now the
specific thresholders can be accessed as class methods, and their arguments
passed as positional arguments, instead.
3. **More expressive (and consistent) transformations.** There's a lot of
ambiguity right now about which methods use which threshold methods by default,
which ones remove self-loops, etc. Under this system, `fit()` just estimates the
weight matrix; it does no thresholding, or binarization, or self-loop removal.
These are handled explicitly by the user through method chaining. So, for
example,

```python
G = (R.fit(TS)
      .remove_self_loops()
      .threshold_on_degree(16)
      .binarize()
      .to_graph())
```

This order of operations also removes ambiguity about the relationship between
`threshold_on_degree()` and `remove_self_loops()`; i.e., should those count
towards the thresholding or not? Now, you can decide for yourself by reversing
the order.
4. **More flexibility in return types.** For a lot of experiments, we didn't
actually need the graphs, but we needed the underlying matrix and had to finish
around `self.results` for the right matrix to use. This got easier after we
standardized on `self.results['weights_matrix']` but was still pretty
inaccessible. Now, the return type is made flexible through the addition of the
`.to_graph()` and `.to_matrix()` methods.

There are some notable downsides to go with the benefits, however.

1. **This approach has a lot of statefulness.** Let's say I'm working in
Jupyter, using a slow reconstruction method. I create my slow reconstructor

```python
R = netrd.reconstruction.SlowReconstructor()
R = R.fit(TS)
```

I'm not sure what thresholding to use, so I play around a bit.

```python
R = R.threshold_on_degree(4)
R = R.threshold_on_degree(16)
```

This isn't going to work, because the first `threshold_on_degree()` has already
chucked out a lot of the information from the weights matrix generated by the
`fit()` call. This is the big problem here; I think it's a pretty serious
weakness but not necessarily a fatal one. If there are ways to avoid unexpected
statefulness, though, I'm all ears.
2. **Possible memory inefficiency.** The internal `BaseReconstructor` methods
are copying dense matrices around a lot, which is possible memory hog. I don't
think this is too big of a deal because the time complexity of most of the
reconstruction algorithms limit their use to small-to-moderate graph sizes
anyway.