remove update_graph, add sparse matrix operations
This commit does not address any of the statefulness concerns. It focuses on the
issues of memory usage and the canonical representation of the reconstructed
network. To the former, it uses sparse matrices to represent thresholded graphs.
To the latter, it removes `update_graph()` and some of the conditionals; the
idea is that the following workflow is implied:

1. The reconstructor is initialized with `self.matrix` and `self.graph` as
`None`.
2. The `fit()` method creates a matrix of weights that resides at `self.matrix`.
3. Optionally, the matrix is thresholded, turning `self.matrix` from a numpy
array to a scipy sparse matrix.
4. Optionally, other operations are performed, like the removal of self loops or
binarization. These functions must work with both dense and sparse matrices.
5. The graph is accessed through the `to_graph()` method, which creates the
networkx object from `self.matrix`. This graph is stored at `self.graph` and
returned again if `to_graph()` is called again.