make copies in non-fit BaseReconstructor methods
All non-`fit()` methods that return `self` now operate on copies of the object.
This should receive test coverage.

Trying it out interactively, it seems to work as expected, returning new copies
of things where appropriate. Memory usage didn't appreciably increase,
presumably because the threshold functions are creating sparse matrices, but we
should revisit when matrices are being copied or not if the whole class is being
copied too.

For now I've just repeated the relevant lines of code. In a general cleanup it
would probably make sense to use a decorator for this, probably just copying
from the MIT-licensed SQLAlchemy repo.