Add test coverage for invariance under vertex relabeling
This adds a test to `test_distance` to assert that the distance between a graph
and a copy of that graph with its node labels permuted should be zero. This is
not something we are strictly enforcing as a requirement but we should have test
coverage for it for documentation purposes; i.e., if we have a distance that is
not invariant here it should be documented, and by enforcing it in test coverage
we hopefully guarantee that the ones we expect to be invariant are. In the
current commit I've excluded all the distances for which the test failed. Some
of these, like `Hamming` and `JaccardDistance`, were unsurprising to me, but
others were. We should make sure that this list represents our prior knowledge
before merging.