Implement graph diffusion distance (#293)
* Implement graph diffusion distance

This PR adds graph diffusion distance, which we were using the in the WEGD
paper. Don't insta-merge; I think there are some variable names, etc. to be
cleaned up, especially in the helper function `gdd_xi_t()`, but I wanted to get
the ball rolling on getting this included.

* changes

Make the recommended changes to the PR. Refer to the PR discussion for more context.

* correct docstring for exponential_diffusion_diff

* refactor and clarify confusing matrix multiplication

* autoformatter

* more prominent credit to authors, plus relicense

* sort eigenvalues

* add gdd to EXCLUDED_DISTANCES

* autoformat